pyproject.toml
[project]
name = "llm-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-google-genai>=3.2.0",
    "langchain-google-vertexai>=3.1.1",
    "langchain[anthropic]>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
